Added setup script, refactored and cleaned code & directory structure
