--- newfile.txt
@@ -1,1 +1,0 @@
-Make certain you do not have a blank repo
