+++ system/display/index.html
@@ -29,7 +29,7 @@
 	  -->
         </form>
 
-        <script src = "main.js"></script>
+        <script src = "../main.js"></script>
       </div>
     </div>
     <div id="cashierwindow" style ="display:none;">
