+++ system/package.json
@@ -11,7 +11,7 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "electron": "^6.0.12",
+    "electron": "^6.1.2",
     "electron-builder": "^21.2.0",
     "mocha": "^6.2.0",
     "spectron": "^8.0.0"
